holo-files: Move FileBuffer from "impl" to "common".
